v4.0.22
Fixed so DebugLevel defaults to .DEBUG and if a message is set to .SILENT it will never print.  Added .SILENT level.  Added colors.  Added emoji for color in console.  Added basic colors to KuColor.  Added rotate() function to DebugLevel for testing and made CaseIterable.  Added array sorting by KeyPath.  Added DebugLevel.colorLogging to KuColor methods so that parsing a string that fails (which might be intentional) doesn't log unnecessarily.