v4.0.24
Updated Application.main.name to include a parenthetical if running in Playground or #Preview.  Added in better loggging level to Application.track() to be clearer where tracked.  Added string literal initialization and assignment of Version.  Removed unnecessary rotate() function since we already have that in ++ function on CaseIterable (made sure that was public as well).